requirements/netprobify.txt
ipam-client==0.6.2
Flask
flask-httpauth
kubernetes==29.0.0
prometheus_client
pykwalify
PyYAML
requests
setuptools
scapy==2.5.0
tornado
waitress
